chore(report-management): extract the `getReportedEvaluation` to the `PlayerController` to use the service
Spring can't inject the Service dependencies in an object class.
